# feat(crypto): use minio/sha256-simd