--- a/go.mod
+++ b/go.mod
@@ -12,6 +12,7 @@ require (
 	github.com/gotd/xor v0.1.1
 	github.com/k0kubun/pp/v3 v3.0.7
 	github.com/klauspost/compress v1.13.1
+	github.com/minio/sha256-simd v1.0.0
 	github.com/pkg/errors v0.9.1 // indirect
 	github.com/quasilyte/go-ruleguard/dsl v0.3.6
 	github.com/rogpeppe/go-internal v1.8.0