```
name             old time/op    new time/op    delta
Read/16b-16         749ns ± 0%     659ns ± 2%  -11.91%
Read/128b-16        830ns ± 1%     740ns ± 4%  -10.92%
Read/1024b-16      1.43µs ± 1%    1.34µs ± 8%   -5.96%
Read/8192b-16      5.26µs ± 9%    5.53µs ±16%   +4.99%
Read/65536b-16     26.9µs ± 2%    33.0µs ±14%  +22.49%
Read/524288b-16     139µs ± 3%      89µs ± 1%  -36.00%

name             old speed      new speed      delta
Read/16b-16      21.4MB/s ± 0%  24.3MB/s ± 2%  +13.53%
Read/128b-16      154MB/s ± 1%   173MB/s ± 4%  +12.33%
Read/1024b-16     717MB/s ± 1%   766MB/s ± 8%   +6.80%
Read/8192b-16    1.57GB/s ± 8%  1.51GB/s ±18%   -3.98%
Read/65536b-16   2.43GB/s ± 2%  2.01GB/s ±15%  -17.46%
Read/524288b-16  3.78GB/s ± 3%  5.90GB/s ± 1%  +56.16%

name             old alloc/op   new alloc/op   delta
Read/16b-16        1.37kB ± 0%    1.37kB ± 0%    0.00%
Read/128b-16       1.59kB ± 0%    1.59kB ± 0%    0.00%
Read/1024b-16      3.48kB ± 0%    3.48kB ± 0%    0.00%
Read/8192b-16      20.1kB ± 0%    20.1kB ± 0%   -0.00%
Read/65536b-16      148kB ± 0%     148kB ± 1%   +0.03%
Read/524288b-16    1.07MB ± 0%    1.07MB ± 0%   +0.10%

name             old allocs/op  new allocs/op  delta
Read/16b-16          16.0 ± 0%      15.7 ± 4%   -2.08%
Read/128b-16         16.0 ± 0%      16.0 ± 0%    0.00%
Read/1024b-16        16.0 ± 0%      16.0 ± 0%    0.00%
Read/8192b-16        16.0 ± 0%      16.0 ± 0%    0.00%
Read/65536b-16       15.5 ± 3%      15.7 ± 4%   +1.08%
Read/524288b-16      16.0 ± 0%      18.3 ± 4%  +14.58%
```

New benchmark is too noisy, should investigate before merge.